completata esercitazione

+++ app/Http/Controllers/ArticlesController.php
@@ -14,18 +14,34 @@ class ArticlesController extends Controller
 
 
     private $articles;
-
+    private const SUPER_USER = 1;
 
     public function __construct(){
         $this->articles = Article::all();
     }
 
+    private function isSuperUser(){
+        return auth()->user()->id === self::SUPER_USER;
+    }
+
     public function index(){
-        return view("articles.index",["title"=> "Articoli", "articles" => $this->articles]);
+
+        if($this->isSuperUser()){
+            $articles = Article::all();
+        }else {
+            $articles = auth()->user()->articles;
+        }
+
+        return view("articles.index",["title"=> "Articoli", "articles" => $articles]);
+    }
+
+    public function articles(){
+        
+        return view('articles.articles',
+        ["title" => "articoli","articles" => $this->articles]);
     }
 
     public function create(){
-
 
         return view("articles.create", [
             "categories" => Category::all(),
@@ -42,16 +58,18 @@ class ArticlesController extends Controller
         //     auth()->user()->email
         // ));
 
+        $article->user_id = auth()->user()->id;
+
         if($request->hasFile("image") && $request->file("image")->isValid()){
             $imgName = "img_" . uniqid() . "." . $request->file("image")->extension();
             $imgPath = "public/images/" . $article->id;
             
 
-            $article->image = $request->file("image")->storeAs($imgPath,$imgName);;
-            $article->save();
+            $article->image = $request->file("image")->storeAs($imgPath,$imgName);
             
         }
-
+        
+        $article->save();
         return redirect()->back()->with("success","Articolo creato correttamente");
     }
 
@@ -61,23 +79,23 @@ class ArticlesController extends Controller
 
 
     public function edit(Article $article){
-        return  view("articles.edit",["title"=> "Modifica Articolo","article"=> $article, "categories" => Category::all()]);
-    }
-
-    public function articles(){
         
-        return view('articles.articles',
-        ["title" => "articoli","articles" => $this->articles]);
+        if($article->user_id === auth()->user()->id || $this->isSuperUser()){
+            return  view("articles.edit",["title"=> "Modifica Articolo","article"=> $article, "categories" => Category::all()]);
+        }
+        
+        abort(403);
     }
-
- 
 
 
     public function update(StoreArticleRequest $request, Article $article)
     {
-        $article->update($request->all());
 
-        return redirect()->route("articles.index")->with("success","Articolo modificato con successo");
+        if($article->user_id === auth()->user()->id || $this->isSuperUser()){
+            $article->update($request->all());
+            return redirect()->route("articles.index")->with("success","Articolo modificato con successo");
+        }
+        abort(403);    
     }
 
     /**
@@ -85,9 +103,13 @@ class ArticlesController extends Controller
      */
     public function destroy(Article $article)
     {
-        $article->delete();
 
-        return redirect()->back();
+        if($article->user_id === auth()->user()->id || $this->isSuperUser()){
+            $article->delete();
+            return redirect()->back()->with(["success" => "Articolo eliminato correttamente"]);
+        }
+        abort(403); 
+        
     }
 
 }
